refactor: make plugin impl more informative
